requirements.txt
stem>=1.8.0
requests>=2.31.0
PySocks>=1.7.1
rich>=13.3.0
